Split DocumentType out into its own module
