+++ medicines/api/src/schema.rs
@@ -2,7 +2,8 @@ use juniper::{FieldResult, RootNode};
 
 use crate::{
     azure_context::AzureContext,
-    document::{get_documents, DocumentType, Documents},
+    document::{get_documents, Documents},
+    document_type::DocumentType,
     product::{get_product, get_substance_with_products, Product},
     substance::{get_substances_starting_with_letter, Substance},
 };
